pyproject.toml
[project]
name = "jira-loading-monitor"
version = "0.1.0"
description = "Jira Loading Progress Monitor using CrewAI with Atlassian MCP"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.5.0",
    "python-dotenv==1.0.0",
    "atlassian-python-api==3.41.8",
    "requests==2.31.0"
]

[project.scripts]
jira_monitor = "jira_loading_monitor.main:run"
run_monitor = "jira_loading_monitor.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

